per the feedback, created an article showcard and passed data as props to the ArticleShowCard component
